#5560 move static methods to PlotRange
