+++ js/notebook/src/plot/range/PlotRange.ts
@@ -17,6 +17,7 @@
 import * as d3 from 'd3';
 import {ScaleLinear} from "d3";
 import PlotFocus from "../zoom/PlotFocus";
+import PlotAxisFactory from "../std/axis/PlotAxisFactory";
 
 const plotUtils = require('../plotUtils');
 
@@ -42,6 +43,54 @@ export default class PlotRange {
 
   constructor(scope) {
     this.scope = scope;
+  }
+
+  static updateAxisXRange(xAxis, model) {
+    const vrange = model.vrange;
+    const xAxisLabel = model.xAxis.label;
+
+    if (xAxis.axisType === "category") {
+      xAxis.setRange(vrange.xl, vrange.xr, model.xAxis.base);
+      xAxis.setCategoryNames(model.categoryNames, model.labelsxs);
+    } else if (xAxis.axisType === "time" || xAxis.axisType === "nanotime") {
+      xAxis.setRange(vrange.xl, vrange.xr, model.timezone);
+    } else {
+      xAxis.setRange(vrange.xl, vrange.xr, model.xAxis.base);
+    }
+
+    if (xAxisLabel != null) {
+      xAxis.setLabel(xAxisLabel);
+    }
+
+    return xAxis;
+  }
+
+  static updateAxisYRange(modelAxis, axisVRange, model) {
+    if (modelAxis == null || axisVRange == null) {
+      return null;
+    }
+
+    const vrange = model.vrange;
+    const label = modelAxis.label;
+    const axis = PlotAxisFactory.getPlotAxis(modelAxis.type);
+
+    if (axis.axisType === "category") {
+      axis.setRange(vrange.xl, vrange.xr, model.xAxis.base);
+      axis.setCategoryNames(model.categoryNames, model.labelsxs);
+    } else if (axis.axisType !== "time") {
+      axis.setRange(axisVRange.yl, axisVRange.yr, modelAxis.base);
+    } else {
+      axis.setRange(axisVRange.yl, axisVRange.yr, modelAxis.timezone);
+    }
+
+    if (label != null) {
+      axis.setLabel(label);
+    }
+
+    axis.axisMarginValL = modelAxis.lowerMargin;
+    axis.axisMarginValR = modelAxis.upperMargin;
+
+    return axis;
   }
 
   calcRange() {
